Move more 2d funcs to geometry module.
