+++ samples/cpp/snippets/segment_objects.cpp
@@ -1,4 +1,5 @@
 #include "opencv2/imgproc.hpp"
+#include "opencv2/geometry.hpp"
 #include "opencv2/videoio.hpp"
 #include "opencv2/highgui.hpp"
 #include "opencv2/video/background_segm.hpp"
