+++ samples/cpp/tutorial_code/ShapeDescriptors/moments_demo.cpp
@@ -7,6 +7,7 @@
 #include "opencv2/imgcodecs.hpp"
 #include "opencv2/highgui.hpp"
 #include "opencv2/imgproc.hpp"
+#include "opencv2/geometry.hpp"
 #include <iostream>
 #include <iomanip>
