+++ samples/cpp/tutorial_code/ml/introduction_to_pca/introduction_to_pca.cpp
@@ -6,6 +6,7 @@
 
 #include "opencv2/core.hpp"
 #include "opencv2/imgproc.hpp"
+#include "opencv2/geometry.hpp"
 #include "opencv2/highgui.hpp"
 #include <iostream>
